requirements.txt
anyio==3.6.2; python_full_version >= "3.6.2" and python_version >= "3.7"
certifi==2022.12.7; python_version >= "3.7"
commonmark==0.9.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0"
h11==0.14.0; python_version >= "3.7"
httpcore==0.16.2; python_version >= "3.7"
httpx==0.23.1; python_version >= "3.7"
idna==3.4; python_version >= "3.7" and python_full_version >= "3.6.2"
pygments==2.13.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.6"
python-dotenv==0.21.0; python_version >= "3.7"
rfc3986==1.5.0; python_version >= "3.7"
rich==12.6.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
